# backend/packages/users-service/openapi.yaml
openapi: 3.1.0
info:
  title: Users Service API
  version: 1.0.0
  description: |
    Serviço de gestão de usuários para o sistema de patrimônio.
    Fornece endpoints para CRUD completo de usuários, busca avançada, validação de credenciais e estatísticas.
servers:
  - url: http://localhost:3002
    description: Servidor de desenvolvimento local
  - url: http://users-service:3002
    description: Servidor Docker interno
paths:
  /users:
    get:
      summary: Listar usuários com filtros avançados
      description: |
        Lista todos os usuários com paginação e filtros avançados.
        Requer autenticação e role MANAGER ou ADMIN.
      operationId: listUsers
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: q
          in: query
          schema:
            type: string
          description: Busca textual (nome e email)
        - name: role
          in: query
          schema:
            type: string
            enum: [OPERATOR, MANAGER, ADMIN]
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, email, createdAt, updatedAt]
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsersResponse'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Criar novo usuário
      description: |
        Cria um novo usuário no sistema.
        Requer autenticação e role ADMIN.
      operationId: createUser
      tags: [users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email já existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    get:
      summary: Buscar usuário por ID
      description: |
        Busca um usuário pelo ID.
        Requer autenticação. O usuário pode buscar apenas seus próprios dados, exceto ADMIN que pode buscar qualquer usuário.
      operationId: getUserById
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: ID inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Atualizar usuário
      description: |
        Atualiza um usuário existente.
        Requer autenticação. O usuário pode atualizar apenas seus próprios dados, exceto ADMIN que pode atualizar qualquer usuário.
      operationId: updateUser
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: Usuário atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Deletar usuário
      description: |
        Deleta um usuário (soft delete).
        Requer autenticação e role ADMIN.
      operationId: deleteUser
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Usuário deletado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User deleted successfully
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/email/{email}:
    get:
      summary: Buscar usuário por email
      description: |
        Busca um usuário pelo email.
        Requer autenticação.
      operationId: getUserByEmail
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Email inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/validate:
    post:
      summary: Validar credenciais de usuário
      description: |
        Valida as credenciais (email e senha) de um usuário.
        Endpoint público usado pelo Auth Service.
        Retorna o usuário se as credenciais forem válidas, null caso contrário.
      operationId: validateCredentials
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateUserRequest'
      responses:
        '200':
          description: Credenciais válidas ou inválidas
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/User'
                  - type: 'null'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/bulk:
    post:
      summary: Criar múltiplos usuários
      description: |
        Cria múltiplos usuários de uma vez.
        Requer autenticação e role ADMIN.
      operationId: createBulkUsers
      tags: [users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                users:
                  type: array
                  items:
                    $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: Usuários criados
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: integer
                  failed:
                    type: integer
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/advanced/search:
    get:
      summary: Busca avançada de usuários
      description: |
        Realiza busca avançada com múltiplos filtros.
        Requer autenticação e role MANAGER ou ADMIN.
      operationId: advancedSearch
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
            enum: [OPERATOR, MANAGER, ADMIN]
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Resultados da busca
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/stats/roles:
    get:
      summary: Estatísticas de usuários por role
      description: |
        Retorna estatísticas de quantos usuários existem por role.
        Requer autenticação e role MANAGER ou ADMIN.
      operationId: getStatsByRoles
      tags: [users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Estatísticas por role
          content:
            application/json:
              schema:
                type: object
                properties:
                  OPERATOR:
                    type: integer
                  MANAGER:
                    type: integer
                  ADMIN:
                    type: integer
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    UpdateUser:
      type: object
      properties:
        name:
          type: string
          example: João Silva
        email:
          type: string
          format: email
          example: joao.silva@example.com
        password:
          type: string
          format: password
          minLength: 8
        role:
          type: string
          enum: [OPERATOR, MANAGER, ADMIN]
        isActive:
          type: boolean
        avatarUrl:
          type: string
          format: uri
          nullable: true
...
    PaginatedUsersResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        meta:
          type: object
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
            total:
              type: integer
              example: 100
            totalPages:
              type: integer
              example: 10
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Validation failed
        error:
          type: string
          example: Bad Request
        code:
          type: string
          example: VALIDATION_ERROR
          nullable: true
        details:
          type: array
          items:
            type: string
          nullable: true
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtido no endpoint /auth/login do Auth Service